requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
beautifulsoup4==4.14.3
boto3==1.42.49
botocore==1.42.49
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
Django==5.0.14
django-ninja==1.5.3
idna==3.11
jmespath==1.1.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
gunicorn==24.1.1
